refactor: add responsive for tablet and desktop, and small fixes
This is a big commit because some small changes where done at the same time the global containers where added. It would be too much useless work to split it in many commits, and it's not a big deal as this is a big refactor that won't be reverted.

Add global containers in components.
Add a desktop max width used by these containers.
Add a breakpoint to deal with an intermediate size.
Add responsive for tablet and desktop.
Add mailto for contact.
Move the favorite icon outside the "a" link component in cards, and place it with absolute position. This is done to avoid clicking on the link when actually clicking on the icon. Playing with z-index only wasn't sufficient.