refactor: delegate cards width property to the parent container
In the case of a cards' mosaic, it's more elegant and readable that the parent container handle the width of his children, rather than the children themselves. A child component shouldn't handle his placement in the container. It's more readable because this is also where we define the gap between rows in percentage, which depends on this width.
Now all the placement of the children is done by the container, using the children selector.